Return whether a mismatch indicates a problem
Previously we only output an error message. This now allows us to check
the return value by examining the return of the function running the
heuristics. This is to allow writing tests.

I also rename the checksum mismatch function. I want to clarify that
calling this function does not mean there is a problem.

I also improved the comments in the check function.

This includes unit tests for checksum_mismatch().