Move config validation to separate function
We call this validation after reading the config. This means we no
longer need to do so when calling the top level check function, which
allows that function to be more of a simple wrapper.

This also reformats some of the tests, and is more careful about
cleaning up temporary files.