Restrict the query for retrieving checksums
As it was, it was possible to load checksums for paths we would not be
looking at. This would have no impact other than waste memory.

For example, if we had top level path /test and /test2, then previously
the query for /test checksums would retrieve those under /test2 as well,
as the query was 'LIKE /test%'. Now we ensure it only includes files
./test/%'.