Use last checksum time in reporting heuristic instead of 24 hours ago
This makes the heuristic more intelligent. Previously the heuristic
assumed any changes to the file within the last 24 hours was okay, and
any before that was problematic. Now we treat any prior (by modified
time) to the last time we computed the checksum as a problem.

This works by us now always storing the time just as the checksum was
computed for each file. There are two downsides to this. First, we have
to constantly get the current time. Second, we now have to update the
database for every file each time we run, even if the file is not new or
its checksum did not change. Whether this will be too onerous on runtime
I've yet to see.

It does give us a benefit though: We can prune records from the database
for files we don't see on the current run. This will mean both that the
database will be smaller, and also that we won't report false positive
corruption issues if the files were moved away, and then moved back.
Although of course the file will have to have been changed and its
modified time set back prior to the time we computed the checksum, so
this seems unlikely.