Compare modified times when checksums mismatch
Previously I compared the current modified time against the last
time the checksum was computed. This lead to false positives. Consider
the case where we compute a checksum for a file, then a backup runs
(where modified times for files are prior to when we computed the
checksum). We'd see legitimate checksum changes, but since the modified
time was before the last time we computed a checksum, the heuristic
would incorrectly say that the modification was suspicious. Now we
compare only against the two modified times.

This means we record the file's modified time at the time we inspect it
in each scan.